no packages in distributions config
By defining attributes like `src` and `package` in distributions config
means that it's not possible evaluate the configuration without also
fetching all sources and building packages.

A future version may reintroduce the packages in a separate top-level
configuration attribute, but I don't think they should be declared in
the same place as the original config.